Silencing uv warnings in CI.
There is a problem when running `uv sync` or `uv run` while the
`turbopelican` virtual environment is activated, which means uv will
warn the developer that one virtual environment is activated while the
virtual environment being touched is different. This should silence the
warning, as the intended behaviour is to ignore the activated virtual
environment.